Efficiency changes, Accel mag, time based trails
cleaned up code a bit in all files. Added an acceleration angle and
magnitude value to the physicsObject class. This can be used to
illistrate the direction and magnitude of the objects velocity. Trails
are now recorded based on a time difference instead of every frame.
Trails are also no longer cleared when turned off, they are always
recorded just not shown unless the drawTrails boolean is true.