Added Object base class declairation
Added a new class called Object. This class will act as the base class
for any object in the World. It contains all(i think!) of the parameters
that an object realisticaly needs in general. This base class is made to
enhance the physics object class, and later the static object class.