Added shape object, World now draws.
Added a shape object to the Object class. This eliminates the setSize()
method from the equaion as the size is set on object creaton ONLY and at
no other time unless width or height is changed.  On top of eliminating
the setSize() method, this makes the object class significantly bigger,
meaning I may need to declair the objects on the heap instead of on the
stack. Also, the world now receives a refrence to the render window to
draw the objects it contains to the window.