Made SFML path for physics point SFML in repo
made sfml in physics point to SFMl in the repo. This makes all programs
useable on both desktop and laptop without changing the paths, as long
as i have changed the paths to the new ones. Also some minor efficiency
tweaks. Objects and their acceleration magnitudes will always disapear
when off screen and reapear when on screen.